Creation of the basics of the project
Created a class State and modified the class Automata to adapt it.

Created a file Automata.py for the classes and another usage.py for other function.
The main.py is now more clear

Function created by Remi to create the table have been modified to work with this new base